Move listener tests to the corresponding conformance test class
