+++ src/IceRpc/Router.cs
@@ -1,7 +1,6 @@
 // Copyright (c) ZeroC, Inc. All rights reserved.
 
 using IceRpc.Slice;
-using System.Diagnostics;
 
 namespace IceRpc;
