+++ tests/IceRpc.Conformance.Tests/Transports/DuplexConnectionConformanceTests.cs
@@ -61,54 +61,6 @@ public abstract class DuplexConnectionConformanceTests
         clientConnection.Dispose();
     }
 
-    /// <summary>Verifies that connect fails if the listener is disposed.</summary>
-    [Test]
-    public async Task Connect_fails_if_listener_is_disposed()
-    {
-        // Arrange
-        await using ServiceProvider provider = CreateServiceCollection().BuildServiceProvider(validateScopes: true);
-        IListener<IDuplexConnection> listener = provider.GetRequiredService<IListener<IDuplexConnection>>();
-        var clientTransport = provider.GetRequiredService<IDuplexClientTransport>();
-
-        // A duplex transport listener might use a backlog to accept client connections (e.g.: TCP). So we need to
-        // create and establish client connections until a connection establishment blocks to test cancellation.
-        Task<TransportConnectionInformation> connectTask;
-        IDuplexConnection clientConnection;
-        while (true)
-        {
-            IDuplexConnection? connection = clientTransport.CreateConnection(
-                listener.ServerAddress,
-                provider.GetService<IOptions<DuplexConnectionOptions>>()?.Value ?? new(),
-                clientAuthenticationOptions: provider.GetService<SslClientAuthenticationOptions>());
-            try
-            {
-                connectTask = connection.ConnectAsync(default);
-                await Task.Delay(TimeSpan.FromMilliseconds(100));
-                if (connectTask.IsCompleted)
-                {
-                    await connectTask;
-                }
-                else
-                {
-                    clientConnection = connection;
-                    connection = null;
-                    break;
-                }
-            }
-            finally
-            {
-                connection?.Dispose();
-            }
-        }
-
-        // Act
-        await listener.DisposeAsync();
-
-        // Assert
-        Assert.That(async () => await connectTask, Throws.InstanceOf<IceRpcException>());
-        clientConnection.Dispose();
-    }
-
     /// <summary>Write data until the transport flow control starts blocking, at this point we start a read task and
     /// ensure that this unblocks the pending write calls.</summary>
     [Test]
@@ -152,20 +104,6 @@ public abstract class DuplexConnectionConformanceTests
                 size -= await connection.ReadAsync(buffer, default);
             }
         }
-    }
-
-    [Test]
-    public async Task Listen_twice_on_the_same_address_fails_with_a_transport_exception()
-    {
-        // Arrange
-        await using ServiceProvider provider = CreateServiceCollection().BuildServiceProvider(validateScopes: true);
-        var listener = provider.GetRequiredService<IListener<IDuplexConnection>>();
-        var serverTransport = provider.GetRequiredService<IDuplexServerTransport>();
-
-        // Act/Assert
-        IceRpcException? exception = Assert.Throws<IceRpcException>(
-            () => serverTransport.Listen(listener.ServerAddress, new DuplexConnectionOptions(), null));
-        Assert.That(exception!.IceRpcError, Is.EqualTo(IceRpcError.AddressInUse));
     }
 
     [Test]
@@ -285,18 +223,6 @@ public abstract class DuplexConnectionConformanceTests
         // Assert
         Assert.That(sut.ClientConnection.ServerAddress.Transport, Is.EqualTo(transport));
         Assert.That(sut.ServerConnection.ServerAddress.Transport, Is.EqualTo(transport));
-    }
-
-    [Test]
-    public async Task Listener_server_address_transport_property_is_set()
-    {
-        // Arrange
-        await using ServiceProvider provider = CreateServiceCollection().BuildServiceProvider(validateScopes: true);
-        var transport = provider.GetRequiredService<IDuplexClientTransport>().Name;
-        var listener = provider.GetRequiredService<IListener<IDuplexConnection>>();
-
-        // Act/Assert
-        Assert.That(listener.ServerAddress.Transport, Is.EqualTo(transport));
     }
 
     [Test]
