+++ tests/IceRpc.Conformance.Tests/Transports/MultiplexedListenerConformanceTests.cs
@@ -2,7 +2,10 @@
 
 using IceRpc.Transports;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Options;
 using NUnit.Framework;
+using System.Net.Security;
+using System.Security.Authentication;
 
 namespace IceRpc.Conformance.Tests;
 
@@ -90,6 +93,88 @@ public abstract class MultiplexedListenerConformanceTests
         Assert.That(
             async () => await listener.AcceptAsync(new CancellationToken(canceled: true)),
             Throws.InstanceOf<OperationCanceledException>());
+    }
+
+    /// <summary>Verifies that connect fails if the listener is disposed.</summary>
+    [Test]
+    public async Task Connect_fails_if_listener_is_disposed()
+    {
+        // Arrange
+        await using ServiceProvider provider = CreateServiceCollection().BuildServiceProvider(validateScopes: true);
+        IListener<IMultiplexedConnection> listener = provider.GetRequiredService<IListener<IMultiplexedConnection>>();
+        var clientTransport = provider.GetRequiredService<IMultiplexedClientTransport>();
+
+        Task connectTask = ConnectAsync(clientTransport);
+
+        // Act
+        await listener.DisposeAsync();
+
+        // Assert
+
+        // If using Quic and the listener is disposed during the ssl handshake this can fail
+        // with AuthenticationException otherwise it fails with TransportException.
+        Assert.That(
+            async () => await connectTask,
+            Throws.InstanceOf<IceRpcException>().Or.TypeOf<AuthenticationException>());
+
+        async Task ConnectAsync(IMultiplexedClientTransport clientTransport)
+        {
+            // Establish connections until we get a failure.
+            var connections = new List<IMultiplexedConnection>();
+            try
+            {
+                while (true)
+                {
+                    IMultiplexedConnection connection = clientTransport.CreateConnection(
+                        listener.ServerAddress,
+                        provider.GetRequiredService<IOptions<MultiplexedConnectionOptions>>().Value,
+                        provider.GetService<SslClientAuthenticationOptions>());
+                    connections.Add(connection);
+
+                    await connection.ConnectAsync(default);
+
+                    // Continue until connect fails.
+                }
+            }
+            finally
+            {
+                await Task.WhenAll(connections.Select(c => c.DisposeAsync().AsTask()));
+            }
+        }
+    }
+
+    [Test]
+    public async Task Listen_twice_on_the_same_address_fails_with_a_transport_exception()
+    {
+        // Arrange
+        await using ServiceProvider provider = CreateServiceCollection().BuildServiceProvider(validateScopes: true);
+        var listener = provider.GetRequiredService<IListener<IMultiplexedConnection>>();
+        var serverTransport = provider.GetRequiredService<IMultiplexedServerTransport>();
+
+        // Act/Assert
+        IceRpcException? exception = Assert.Throws<IceRpcException>(
+            () => serverTransport.Listen(
+                listener.ServerAddress,
+                new MultiplexedConnectionOptions(),
+                provider.GetService<SslServerAuthenticationOptions>()));
+        // BUGFIX with Quic this throws an internal error https://github.com/dotnet/runtime/issues/78573
+        Assert.That(
+            exception!.IceRpcError,
+            Is.EqualTo(IceRpcError.AddressInUse).Or.EqualTo(IceRpcError.IceRpcError));
+    }
+
+    [Test]
+    public async Task Listener_server_address_transport_property_is_set()
+    {
+        // Arrange
+        await using ServiceProvider provider = CreateServiceCollection()
+            .AddMultiplexedTransportTest()
+            .BuildServiceProvider(validateScopes: true);
+        var transport = provider.GetRequiredService<IMultiplexedClientTransport>().Name;
+        var listener = provider.GetRequiredService<IListener<IMultiplexedConnection>>();
+
+        // Act/Assert
+        Assert.That(listener.ServerAddress.Transport, Is.EqualTo(transport));
     }
 
     /// <summary>Creates the service collection used for multiplexed listener transport conformance tests.</summary>
