+++ tests/IceRpc.Tests/IceRpcProtocolConnectionTests.cs
@@ -867,7 +867,7 @@ public sealed class IceRpcProtocolConnectionTests
             async () => await sut.Client.InvokeAsync(request),
             Throws.InstanceOf<IceRpcException>().With.Property("IceRpcError").EqualTo(IceRpcError.TruncatedData));
         Assert.That(
-            async() => await tcs.Task,
+            async () => await tcs.Task,
             Is.InstanceOf<IceRpcException>().With.Property("IceRpcError").EqualTo(IceRpcError.LimitExceeded));
     }
